requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-openai>=0.0.5
pandas>=2.1.0
plotly>=5.17.0
python-dotenv>=1.0.0
pydantic>=2.4.2
openai>=1.3.0
SQLAlchemy>=2.0.23
python-multipart>=0.0.6
llama-index>=0.9.3